chore: don't kill current process

+++ packages/dts-plugin/src/core/lib/DtsWorker.ts
@@ -40,7 +40,10 @@ export class DtsWorker {
     const ensureChildProcessExit = () => {
       try {
         const pid = this.rpcWorker.process?.pid;
-        process.kill(pid, 0);
+        const currentPid = process.pid;
+        if (currentPid !== pid) {
+          process.kill(pid, 0);
+        }
       } catch (error) {
         if (isDebugMode()) {
           console.error(error);
